+++ SLP_Ade/config.py
@@ -104,8 +104,8 @@ TRANSPORT_VARIANT_KEYS = [      #TODO: see above
 HOUSETYPE = lpgdata.HouseTypes.HT20_Single_Family_House_no_heating_cooling
 
 #: Custom binary path for LPG. Set to None to use the official release downloaded automatically by the package.
-LPG_BINARY_PATH = "/fast/home/a-tarasenko/SLP_Ade/LoadProfileGenerator/SimEngine2/bin/release/net9.0/linux-x64/publish/SimEngine2"
-# "pylpg/LPG_win" for local testing
+LPG_BINARY_PATH = "C:\\Tarasenko\\GitHub\\LoadProfileGenerator\\SimEngine2\\bin\\release\\net9.0\\win-x64\\publish\\SimEngine2.exe"  # for local Windows dev
+# "/fast/home/a-tarasenko/SLP_Ade/LoadProfileGenerator/SimEngine2/bin/release/net9.0/linux-x64/publish/SimEngine2" for cluster
 
 # Define runs per combination. You can specify:
 # - A dict mapping combo_tag patterns to run counts